+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.1 - 2023-10-25
+
+### Changed
+ 
+-   Add `#[inline(always)]` to Result impl to improve performance.
+-   Add benches to test performance.
+
 ## 1.0.0 - 2022-10-31
 
 ### Changed